Add skeleton for color space changes
Implemented all changes necessary to add a new feature to the Frontend.
This feature does not yet do anything, howeve, it seems to load and
display pdfs correctly. There is a minor bug when multiple pdfs are
added, but this seems to be consistent with other tools/features.

The only code that is strictly necessary to change now should be in
`change-color-space.js` (to add functionality).