Add method to convert text to another color space
With this most of the base functionality of the API for converting a pdf
to another color space should be finished.

The added method `changeColorSpaceTextAndBackground()` reads through the
data of the entire pdf and changes all `rg` and `RG` with their colors
into the correct color space by remaking the pdf data.

Right now this is done after converting images to CMYK, however, it may
be possible to merge the two methods, to make it quicker.