Finilize functionality of color space change API
These changes finilize the functionality of the color space change API.
Requests can now be done from the related tool, and is automatically
downloaded, and seems to work fine.

Important note is that there currently is only one available ICC
profile, and that the program does not expect resources from
`src/main/resources` but instead the root directory, which is not ideal.
This will be needed to be changed before merging to main.

There are also no tests that check the functionality of the code, for
example if the color conversion is done properly.